requirements.txt
selenium
pandas
openpyxl
webdriver-manager
python-dotenv